add method_missing and respond_to? to ArchiveMethods module
method_missing here is checking if we're trying to invoke a finder, if so it passes it on up the chain to AR::B's MM, if not it checks to see if the original class responds to the instance method called if so we build a new object and let it do the work, and finally if none of these conditions are met we call super and are done with it.

in this case respond_to? is just letting us know that we can indeed call the original class' methods on the Archive.

Signed-off-by: t.pickett66 <t.pickett66@gmail.com>