disable mass assign. protection for archives
In rails 3.1/3.2 they have made the default state
for mass assignment in new projects be set to
disabled, if this is set it keeps us from creating
archives correctly w/o application specific code.
To get around this in all versions of rails >=3.1
we're going to just explicitly add attr_accessable
for all attributes.


Signed-off-by: t.pickett66 <t.pickett66@gmail.com>