fix the migration file naming fail
So the timestamp based migration naming got moved
from somewhere else into the actual AR::Migration
class but it has to go through method missing.
What I've done to fix the problem is just move
the original method missing to the end of the
method allowing the original return value to be
what we get back now.

Signed-off-by: t.pickett66 <t.pickett66@gmail.com>